Move common user input into a dedicated manager
